ack queue: don't hang on to 1000 acked items
Items that have been acked should not be kept around needlessly. This
patch changes the behavior of clear_acked_data() to not retain any acked
data (basically make the function do what the name implies). Additionally
a new function shrink_disk_usage() was introduced to allow the owner of
the ack queue instance to force sqlite to vacuum and shrink the file size.

Fixes #78