fix: don't mark empty queue without max size 'full'
In the current implementation, 'full' only checks if the current size is
the same as the maximum, which fails as '0' is also used to mark queues
without any upper limit.

This change modifies the functions to explicitly checks that a max_size
is set to a value bigger than 0. Tests are added accordingly.

This closes #230

Signed-off-by: Stefan Kraus <dev@stefankraus.org>